+++ test/unit/elastic_db_admin/check_status.py
@@ -50,14 +50,14 @@ class ElasticSearchStatus(object):
 
     """
 
-    def __init__(self, node, port, mem, cpu, disk):
+    def __init__(self, hosts, port, mem, cpu, disk):
 
         """Method:  __init__
 
         Description:  Initialization instance of the class.
 
         Arguments:
-            (input) node -> Node name.
+            (input) hosts -> Hosts name.
             (input) port -> Port number.
             (input) mem -> Memory cutoff.
             (input) cpu -> Cpu cutoff.
@@ -65,7 +65,7 @@ class ElasticSearchStatus(object):
 
         """
 
-        self.node = node
+        self.hosts = hosts
         self.port = port
         self.all_err_msg = None
         self.cluster_err_msg = None
@@ -135,7 +135,7 @@ class ElasticSearch(object):
 
         """
 
-        self.node = "nodename"
+        self.hosts = ["nodename1", "nodename2"]
         self.port = 1234
 
 
@@ -147,6 +147,7 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp -> Initialization for unit testing.
+        test_cutoff_args -> Test passing in cutoff arguments.
         test_one_option_cfg -> Test config settings.
         test_one_option_error -> Test one option with error.
         test_incorrect_option -> Test with incorrect option.
@@ -190,6 +191,25 @@ class UnitTest(unittest.TestCase):
         self.cfg2 = cfg2("test")
 
     @mock.patch("elastic_db_admin.elastic_class.ElasticSearchStatus")
+    def test_cutoff_args(self, mock_class):
+
+        """Function:  test_cutoff_args
+
+        Description:  Test passing in cutoff arguments.
+
+        Arguments:
+
+        """
+
+        mock_class.return_value = ElasticSearchStatus(
+            self.es.hosts, self.es.port, self.mem, self.cpu, self.disk)
+
+        self.assertFalse(
+            elastic_db_admin.check_status(
+                self.es, check_call=self.check_call,
+                args_array=self.args_array8, cfg=self.cfg))
+
+    @mock.patch("elastic_db_admin.elastic_class.ElasticSearchStatus")
     def test_one_option_cfg(self, mock_class):
 
         """Function:  test_one_option_cfg
@@ -201,7 +221,7 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_class.return_value = ElasticSearchStatus(
-            self.es.node, self.es.port, self.mem, self.cpu, self.disk)
+            self.es.hosts, self.es.port, self.mem, self.cpu, self.disk)
 
         self.assertFalse(
             elastic_db_admin.check_status(
@@ -219,7 +239,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        es = ElasticSearchStatus(self.es.node, self.es.port, self.mem,
+        es = ElasticSearchStatus(self.es.hosts, self.es.port, self.mem,
                                  self.cpu, self.disk)
         es.mem_err_msg = {"Err": "Error Message"}
         es.cluster_err_msg = {}
@@ -243,7 +263,7 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_class.return_value = ElasticSearchStatus(
-            self.es.node, self.es.port, self.mem, self.cpu, self.disk)
+            self.es.hosts, self.es.port, self.mem, self.cpu, self.disk)
 
         with gen_libs.no_std_out():
             self.assertFalse(
@@ -263,7 +283,7 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_class.return_value = ElasticSearchStatus(
-            self.es.node, self.es.port, self.mem, self.cpu, self.disk)
+            self.es.hosts, self.es.port, self.mem, self.cpu, self.disk)
 
         self.assertFalse(
             elastic_db_admin.check_status(
@@ -282,7 +302,7 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_class.return_value = ElasticSearchStatus(
-            self.es.node, self.es.port, self.mem, self.cpu, self.disk)
+            self.es.hosts, self.es.port, self.mem, self.cpu, self.disk)
 
         self.assertFalse(
             elastic_db_admin.check_status(
@@ -301,7 +321,7 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_class.return_value = ElasticSearchStatus(
-            self.es.node, self.es.port, self.mem, self.cpu, self.disk)
+            self.es.hosts, self.es.port, self.mem, self.cpu, self.disk)
 
         self.assertFalse(
             elastic_db_admin.check_status(
@@ -318,7 +338,7 @@ class UnitTest(unittest.TestCase):
         Arguments:
 
         """
-        es = ElasticSearchStatus(self.es.node, self.es.port, self.mem,
+        es = ElasticSearchStatus(self.es.hosts, self.es.port, self.mem,
                                  self.cpu, self.disk)
         es.all_err_msg = "Error Message"
         mock_class.return_value = es
@@ -341,7 +361,7 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_class.return_value = ElasticSearchStatus(
-            self.es.node, self.es.port, self.mem, self.cpu, self.disk)
+            self.es.hosts, self.es.port, self.mem, self.cpu, self.disk)
 
         self.assertFalse(
             elastic_db_admin.check_status(
@@ -360,7 +380,7 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_class.return_value = ElasticSearchStatus(
-            self.es.node, self.es.port, self.mem, self.cpu, self.disk)
+            self.es.hosts, self.es.port, self.mem, self.cpu, self.disk)
 
         self.assertFalse(
             elastic_db_admin.check_status(